+++ src/main/kotlin/io/kjson/JSONLong.kt
@@ -28,9 +28,9 @@ package io.kjson
 import java.math.BigDecimal
 import java.util.function.IntConsumer
 
-import net.pwall.util.CoIntOutput.outputLong
-import net.pwall.util.CoOutput
-import net.pwall.util.IntOutput
+import io.jstuff.util.IntOutput
+import io.kstuff.util.CoIntOutput.outputLong
+import io.kstuff.util.CoOutput
 
 /**
  * A JSON long integer value (more than 32 bits but not more than 64 bits).