+++ src/test/kotlin/io/kjson/parser/ParserObjectTest.kt
@@ -2,7 +2,7 @@
  * @(#) ParserObjectTest.kt
  *
  * kjson-core  JSON Kotlin core functionality
- * Copyright (c) 2021, 2022, 2023, 2024 Peter Wall
+ * Copyright (c) 2021, 2022, 2023, 2024, 2025 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -92,21 +92,27 @@ class ParserObjectTest {
     }
 
     @Test fun `should throw exception on missing closing brace`() {
-        shouldThrow<ParseException>(MISSING_CLOSING_BRACE) { Parser.parse("""{"first":123""") }.let {
+        shouldThrow<ParseException>(MISSING_CLOSING_BRACE) {
+            Parser.parse("""{"first":123""")
+        }.let {
             it.text shouldBe MISSING_CLOSING_BRACE
             it.pointer shouldBe rootPointer
         }
     }
 
     @Test fun `should throw exception on missing colon`() {
-        shouldThrow<ParseException>(MISSING_COLON) { Parser.parse("""{"first"123}""") }.let {
+        shouldThrow<ParseException>("$MISSING_COLON, at /first") {
+            Parser.parse("""{"first"123}""")
+        }.let {
             it.text shouldBe MISSING_COLON
-            it.pointer shouldBe rootPointer
+            it.pointer shouldBe "/first"
         }
     }
 
     @Test fun `should throw exception on trailing comma`() {
-        shouldThrow<ParseException>(ILLEGAL_KEY) { Parser.parse("""{"first":123,}""") }.let {
+        shouldThrow<ParseException>(ILLEGAL_KEY) {
+            Parser.parse("""{"first":123,}""")
+        }.let {
             it.text shouldBe ILLEGAL_KEY
             it.pointer shouldBe rootPointer
         }
@@ -123,7 +129,9 @@ class ParserObjectTest {
     }
 
     @Test fun `should throw exception on missing quotes around key`() {
-        shouldThrow<ParseException>(ILLEGAL_KEY) { Parser.parse("{first:123}") }.let {
+        shouldThrow<ParseException>(ILLEGAL_KEY) {
+            Parser.parse("{first:123}")
+        }.let {
             it.text shouldBe ILLEGAL_KEY
             it.pointer shouldBe rootPointer
         }
@@ -140,39 +148,41 @@ class ParserObjectTest {
     }
 
     @Test fun `should throw exception on duplicate keys`() {
-        shouldThrow<JSONException>("Duplicate key - first") { Parser.parse("""{"first":123,"first":456}""") }.let {
-            it.text shouldBe "Duplicate key - first"
-            it.key shouldBe ""
+        shouldThrow<ParseException>("Duplicate key, at /first") {
+            Parser.parse("""{"first":123,"first":456}""")
+        }.let {
+            it.text shouldBe "Duplicate key"
+            it.key shouldBe "/first"
         }
     }
 
     @Test fun `should throw exception on duplicate keys with option ERROR`() {
         val options = ParseOptions(JSONObject.DuplicateKeyOption.ERROR)
-        shouldThrow<JSONException>("Duplicate key - first") {
+        shouldThrow<ParseException>("Duplicate key, at /first") {
             Parser.parse("""{"first":123,"first":456}""", options)
         }.let {
-            it.text shouldBe "Duplicate key - first"
-            it.key shouldBe ""
+            it.text shouldBe "Duplicate key"
+            it.key shouldBe "/first"
         }
     }
 
     @Test fun `should throw exception with pointer on duplicate keys with option ERROR`() {
         val options = ParseOptions(JSONObject.DuplicateKeyOption.ERROR)
-        shouldThrow<JSONException>("Duplicate key - first, at /abc") {
+        shouldThrow<ParseException>("Duplicate key, at /abc/first") {
             Parser.parse("""{"abc":{"first":123,"first":456}}""", options)
         }.let {
-            it.text shouldBe "Duplicate key - first"
-            it.key shouldBe "/abc"
+            it.text shouldBe "Duplicate key"
+            it.key shouldBe "/abc/first"
         }
     }
 
     @Test fun `should throw exception on duplicate keys with option CHECK_IDENTICAL and different values`() {
         val options = ParseOptions(JSONObject.DuplicateKeyOption.CHECK_IDENTICAL)
-        shouldThrow<JSONException>("Duplicate key - first") {
+        shouldThrow<ParseException>("Duplicate key, at /first") {
             Parser.parse("""{"first":123,"first":456}""", options)
         }.let {
-            it.text shouldBe "Duplicate key - first"
-            it.key shouldBe ""
+            it.text shouldBe "Duplicate key"
+            it.key shouldBe "/first"
         }
     }
 
@@ -214,9 +224,11 @@ class ParserObjectTest {
     @Test fun `should throw exception on nesting depth exceeded`() {
         val options = ParseOptions(maximumNestingDepth = 50)
         val excessive = """{"a":""".repeat(options.maximumNestingDepth + 1)
-        shouldThrow<ParseException>(MAX_DEPTH_EXCEEDED) { Parser.parse(excessive, options) }.let {
+        shouldThrow<JSONException>(MAX_DEPTH_EXCEEDED) {
+            Parser.parse(excessive, options)
+        }.let {
             it.text shouldBe MAX_DEPTH_EXCEEDED
-            it.pointer shouldBe rootPointer
+            it.key shouldBe null
         }
     }